requirements.txt
django==4.2.16
djangorestframework==3.15.2
python-decouple==3.8
redis==5.0.8
celery==5.4.0
python-telegram-bot==13.14